frontend: refactor task handling into own mod
