+++ yeet-frontend/src/update/mod.rs
@@ -1,6 +1,5 @@
 use std::{cmp::Ordering, path::Path};
 
-use tokio_util::sync::CancellationToken;
 use yeet_buffer::{
     message::BufferMessage,
     model::{ansi::Ansi, Buffer, BufferLine, Mode},
@@ -11,7 +10,7 @@ use yeet_keymap::message::{KeySequence, KeymapMessage, PrintContent};
 use crate::{
     action::Action,
     event::{Envelope, Message, Preview},
-    model::{BufferType, CurrentTask, Model, WindowType},
+    model::{BufferType, Model, WindowType},
 };
 
 use self::{
@@ -61,6 +60,7 @@ mod search;
 mod selection;
 mod settings;
 mod sign;
+mod task;
 pub mod viewport;
 
 const SORT: fn(&BufferLine, &BufferLine) -> Ordering = |a, b| {
@@ -120,9 +120,9 @@ fn update_with_message(model: &mut Model, message: Message) -> Vec<Action> {
         Message::Rerender => Vec::new(),
         Message::Resize(x, y) => vec![Action::Resize(x, y)],
         Message::TaskStarted(identifier, cancellation) => {
-            add_current_task(model, identifier, cancellation)
+            task::add(model, identifier, cancellation)
         }
-        Message::TaskEnded(identifier) => remove_current_task(model, identifier),
+        Message::TaskEnded(identifier) => task::remove(model, identifier),
         Message::ZoxideResult(path) => navigate_to_path(model, path.as_ref()),
     }
 }
@@ -284,54 +284,4 @@ pub fn buffer_type(
         WindowType::Preview => model.files.preview = buffer_type,
         WindowType::Current => unreachable!(),
     };
-}
-
-fn add_current_task(
-    model: &mut Model,
-    identifier: String,
-    cancellation: CancellationToken,
-) -> Vec<Action> {
-    let id = next_id(model);
-
-    if let Some(replaced_task) = model.current_tasks.insert(
-        identifier.clone(),
-        CurrentTask {
-            token: cancellation,
-            id,
-            external_id: identifier,
-        },
-    ) {
-        replaced_task.token.cancel();
-    }
-    Vec::new()
-}
-
-fn next_id(model: &mut Model) -> u16 {
-    let mut next_id = if model.latest_task_id >= 9999 {
-        1
-    } else {
-        model.latest_task_id + 1
-    };
-
-    let mut running_ids: Vec<u16> = model.current_tasks.values().map(|task| task.id).collect();
-    running_ids.sort();
-
-    for id in running_ids {
-        match next_id.cmp(&id) {
-            Ordering::Equal => next_id += 1,
-            Ordering::Greater => break,
-            Ordering::Less => {}
-        }
-    }
-
-    model.latest_task_id = next_id;
-
-    next_id
-}
-
-fn remove_current_task(model: &mut Model, identifier: String) -> Vec<Action> {
-    if let Some(task) = model.current_tasks.remove(&identifier) {
-        task.token.cancel();
-    }
-    Vec::new()
 }
